Refactoring 0.2
