+++ src/app/todo/page.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+const Todo = () => {
+    return (
+        <div>
+            <h1>Todo</h1>
+            {/* Calendar logic and components will go here */}
+        </div>
+    );
+};
+
+export default Todo;
